% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_spark_types.R
\name{identify_spark_types}
\alias{identify_spark_types}
\title{Identify Corresponding Spark Types}
\usage{
identify_spark_types(.data)
}
\arguments{
\item{.data}{\code{data.frame}}
}
\value{
character vector
}
\description{
Identify Corresponding Spark Types
}
