% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{log_if_verbose}
\alias{log_if_verbose}
\title{Log to stdout if set to verbose}
\usage{
log_if_verbose(...)
}
\arguments{
\item{...}{Args passed to message}
}
\value{
None, side-effect only
}
\description{
Log to stdout if set to verbose
}
