% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetr.R
\docType{data}
\name{parquetr}
\alias{parquetr}
\title{R6 Object to wrap Spark connection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
parquetr
}
\description{
R6 Object to wrap Spark connection
}
\examples{
\dontrun{
temp_table <- uuid::UUIDgenerate()
local_spark <- sparkConnection$new()
local_spark$write_parquet(iris, temp_table)
local_spark$read_parquet(temp_table)
local_spark$delete_parquet(temp_table)
}
}
\keyword{datasets}
