% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read a parquet file as an R dataframe via pandas read_parquet}
\usage{
read_parquet(file)
}
\arguments{
\item{path}{The parquet file path}
}
\description{
This function allows you to read a parquet file as an R data frame.
}
\keyword{parquet}
