% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_parquet.R
\name{write_parquet}
\alias{write_parquet}
\title{Writee a parquet file from R dataframe via pandas write_parquet}
\usage{
write_parquet(dd, file)
}
\arguments{
\item{dd}{A data frame}

\item{path}{A character string naming a filepath}
}
\description{
This function allows you to write a parquet file as an R data frame.
}
\keyword{parquet}
