% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetr.R
\docType{class}
\name{Parquetr}
\alias{Parquetr}
\title{R6 Object to wrap Spark connection}
\format{\code{\link{R6Class}} object.}
\usage{
Parquetr
}
\value{
Object of \code{\link{R6Class}} with methods for reading/writing Parquet on S3.
}
\description{
R6 Object to wrap Spark connection
}
\section{Fields}{

\describe{
\item{\code{bucket}}{A R6S3 function that generates a target bucket}

\item{\code{df}}{A data.frame to process}

\item{\code{location}}{\code{character} Of the location relative to the bucket root (allowing for the root prefix of spark-storage) to interact with}
}}

\section{Methods}{

\describe{
  \item{\code{new(bucket)}}{This method is used to create object of this class with \code{bucket} as the target bucket.}
  \item{\code{write_parquet(df, location, mode = "overwrite", ...)}}{This method is used to create object of this class with \code{bucket} as the target bucket.}
}
}

\examples{
\dontrun{
temp_table <- uuid::UUIDgenerate()
local_spark <- sparkConnection$new()
local_spark$write_parquet(iris, temp_table)
local_spark$read_parquet(temp_table)
local_spark$delete_parquet(temp_table)
}
}
\keyword{data}
